Updated benchmarks
